refact: refactor img and span emoji
