+++ projects/chat/src/lib/environments/environment.default.ts
@@ -1,0 +1,13 @@
+import { ProfileImg, ChatImg } from "./environment.model";
+
+export const defaultProfileImages: ProfileImg = {
+  received: 'https://i.imgur.com/JGcirsa.png',
+  sending: 'https://i.imgur.com/9iHRduT.png',
+  sent: 'https://i.imgur.com/OkgTTP7.png'
+};
+
+export const defaultChatImages: ChatImg = {
+  "dark-mode": 'https://i.imgur.com/wA7V6LY.png',
+  "light-mode": 'https://i.imgur.com/xh1BqSg.png',
+  send: 'https://i.imgur.com/CuUJbN6.png'
+};
